pyproject.toml
[tool.poetry]
name = "divedb"
version = "0.1.0"
description = ""
authors = ["Fabio <fabio@stsk.it>"]
license = "GPL3"
readme = "README.md"
packages = [
    { include = "main", from = "src" },
    { include = "drivers/hive", from = "src" },
	   { include = "drivers/mongodb", from = "src" },
	   { include = "drivers/postgresql", from = "src" },
]
[tool.poetry.dependencies]
python = "3.12.9"
pymongo = "3.12"
pyyml = "^0.0.2"
pyqt6 = "^6.7.0"
attrs = "^23.2.0"
psycopg2 = "^2.9.9"
pytest = "^8.3.3"
thrift = "^0.21.0"
wheel = "^0.45.1"
requests = "^2.32.3"
pytest-docker = "^3.2.0"
pyhive = "^0.7.0"
thrift-sasl = "^0.4.3"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
